# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: CommercePulse analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /marketing/summary:
    get:
      operationId: getMarketingSummary
      tags: [marketing]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketingSummary"
  /marketing/campaigns:
    get:
      operationId: listCampaigns
      tags: [marketing]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
  /marketing/by-channel:
    get:
      operationId: getMarketingByChannel
      tags: [marketing]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Spend & revenue per channel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelMetric"
...
  /insights:
    get:
      operationId: getInsights
      tags: [insights]
      summary: AI-powered business insights and recommendations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Insights list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightsResponse"
  /insights/summary:
    get:
      operationId: getInsightsSummary
      tags: [insights]
      summary: Customer, store, and traffic insights
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Live customer, store, and traffic insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightsSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RangeParam:
      in: query
      name: range
      schema:
        type: string
        enum: [7d, 14d, 30d, 90d]
        default: 30d
  schemas:
...
    Metric:
      type: object
      required: [value, deltaPct]
      properties:
        value: { type: number }
        deltaPct: { type: number }
...
    MarketingSummary:
      type: object
      required:
        [
          adSpend,
          adRevenue,
          roas,
          cpa,
          ctr,
          conversions,
          clicks,
          impressions,
        ]
      properties:
        adSpend: { type: number }
        adRevenue: { type: number }
        roas: { type: number }
        cpa: { type: number }
        ctr: { type: number }
        conversions: { type: integer }
        clicks: { type: integer }
        impressions: { type: integer }
    Campaign:
      type: object
      required:
        [
          id,
          name,
          channel,
          spend,
          revenue,
          conversions,
          clicks,
          impressions,
          cpa,
          roas,
          ctr,
        ]
      properties:
        id: { type: string }
        name: { type: string }
        channel: { type: string }
        spend: { type: number }
        revenue: { type: number }
        conversions: { type: integer }
        clicks: { type: integer }
        impressions: { type: integer }
        cpa: { type: number }
        roas: { type: number }
        ctr: { type: number }
    ChannelMetric:
      type: object
      required: [channel, spend, revenue]
      properties:
        channel: { type: string }
        spend: { type: number }
        revenue: { type: number }
...
    Insight:
      type: object
      required: [id, severity, title, description]
      properties:
        id: { type: string }
        severity: { type: string }
        title: { type: string }
        description: { type: string }
        metric: { type: string }
        action: { type: string }
    InsightsResponse:
      type: object
      required: [insights]
      properties:
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
    InsightsSummary:
      type: object
      required: [customer, store, traffic, highlights]
      properties:
        customer:
          type: object
          required: [totalCustomers, newCustomers, repeatRate, averageLifetimeValue]
          properties:
            totalCustomers: { type: integer }
            newCustomers: { $ref: "#/components/schemas/Metric" }
            repeatRate: { $ref: "#/components/schemas/Metric" }
            averageLifetimeValue: { $ref: "#/components/schemas/Metric" }
            topCustomer:
              type: ["object", "null"]
              required: [name, value]
              properties:
                name: { type: string }
                value: { type: number }
        store:
          type: object
          required: [revenue, orders, averageOrderValue, margin]
          properties:
            revenue: { $ref: "#/components/schemas/Metric" }
            orders: { $ref: "#/components/schemas/Metric" }
            averageOrderValue: { $ref: "#/components/schemas/Metric" }
            margin: { type: number }
        traffic:
          type: object
          required: [impressions, clicks, ctr, conversions]
          properties:
            impressions: { $ref: "#/components/schemas/Metric" }
            clicks: { $ref: "#/components/schemas/Metric" }
            ctr: { $ref: "#/components/schemas/Metric" }
            conversions: { $ref: "#/components/schemas/Metric" }
        highlights:
          type: object
          required: [adSets, creatives, repeatCustomers, suggestions]
          properties:
            adSets:
              type: array
              items:
                $ref: "#/components/schemas/PerformanceHighlight"
            creatives:
              type: array
              items:
                $ref: "#/components/schemas/PerformanceHighlight"
            repeatCustomers:
              type: array
              items:
                $ref: "#/components/schemas/RepeatCustomerHighlight"
            suggestions:
              type: array
              items:
                $ref: "#/components/schemas/Insight"
    PerformanceHighlight:
      type: object
      required: [id, name, channel, spend, revenue, conversions, clicks, impressions, cpa, roas, ctr]
      properties:
        id: { type: string }
        name: { type: string }
        channel: { type: string }
        parentName: { type: string }
        spend: { type: number }
        revenue: { type: number }
        conversions: { type: integer }
        clicks: { type: integer }
        impressions: { type: integer }
        cpa: { type: number }
        roas: { type: number }
        ctr: { type: number }
    RepeatCustomerHighlight:
      type: object
      required: [id, name, email, ordersCount, totalSpent, averageOrderValue]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        ordersCount: { type: integer }
        totalSpent: { type: number }
        averageOrderValue: { type: number }
...